+++ changelog.md
@@ -3,6 +3,11 @@ CHANGELOG
 ====================================================================
 ## 2.0.5
 * NEW: Support PirateForms.
+* IMPROVED: Header cover on single product.
+* IMPROVED: WooCommerce header title.
+* FIXED: Title on single post & WooCommerce category.
+* FIXED: Section background parallax effect.
+* FIXED: Back to top button z-index.
 
 ## 2.0.4
 * REMOVED: Remove debug test.